+++ database/migrations/2020_09_26_131932_add_type_fields_to_visitors_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddTypeFieldsToVisitorsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('visitors', function (Blueprint $table) {
+            $table->enum('type', ['beneficiary', 'staff', 'external'])
+                ->default('beneficiary')
+                ->after('last_name');
+            $table->string('organization')
+                ->after('place_of_residence')
+                ->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('visitors', function (Blueprint $table) {
+            $table->dropColumn('organization');
+            $table->dropColumn('type');
+        });
+    }
+}