+++ resources/js/vue-i18n-locales.generated.js
@@ -499,7 +499,8 @@ export default {
             "checkout_everyone": "Alle auschecken",
             "really_checkout_everyone": "Sollen wirklich alle ausgecheckt werden?",
             "everyone_checked_out": "Alle wurden ausgecheckt.",
-            "back": "Zurück"
+            "back": "Zurück",
+            "organization": "Organisation"
         },
         "auth": {
             "failed": "Diese Kombination aus Zugangsdaten wurden nicht in unserer Datenbank gefunden.",
@@ -1161,7 +1162,11 @@ export default {
         },
         "visitors": {
             "visitors": "Besucher",
-            "current_visitors": "Aktuelle Besucher"
+            "current_visitors": "Aktuelle Besucher",
+            "we_have_n_people_visiting_and_m_people_total_today": "Wir haben aktuell <strong>{current}</strong> besucher und heute <strong>{total}</strong> Besucher in total.",
+            "beneficiary": "Leistungsempfänger",
+            "volunteer_staff": "Freiwilliger / Mitarbeiter",
+            "external_visitor": "Externer Besucher"
         },
         "wiki": {
             "wiki": "Wiki",
@@ -1691,7 +1696,8 @@ export default {
             "checkout_everyone": "Checkout everyone",
             "really_checkout_everyone": "Really checkout everyone?",
             "everyone_checked_out": "Everyone has been checked out.",
-            "back": "Back"
+            "back": "Back",
+            "organization": "Organization"
         },
         "auth": {
             "failed": "These credentials do not match our records.",
@@ -2345,7 +2351,11 @@ export default {
         },
         "visitors": {
             "visitors": "Visitors",
-            "current_visitors": "Current Visitors"
+            "current_visitors": "Current Visitors",
+            "we_have_n_people_visiting_and_m_people_total_today": "We are having <strong>{current}</strong> people visiting and <strong>{total}</strong> visitors in total today.",
+            "beneficiary": "Beneficiary",
+            "volunteer_staff": "Volunteer / Staff",
+            "external_visitor": "External visitor"
         },
         "wiki": {
             "wiki": "Wiki",