+++ resources/lang/de/visitors.php
@@ -4,4 +4,7 @@ return [
     'visitors' => 'Besucher',
     'current_visitors' => 'Aktuelle Besucher',
     'we_have_n_people_visiting_and_m_people_total_today' => 'Wir haben aktuell <strong>:current</strong> besucher und heute <strong>:total</strong> Besucher in total.',
+    'beneficiary' => 'Leistungsempfänger',
+    'volunteer_staff' => 'Freiwilliger / Mitarbeiter',
+    'external_visitor' => 'Externer Besucher',
 ];