+++ resources/lang/en/visitors.php
@@ -4,4 +4,7 @@ return [
     'visitors' => 'Visitors',
     'current_visitors' => 'Current Visitors',
     'we_have_n_people_visiting_and_m_people_total_today' => 'We are having <strong>:current</strong> people visiting and <strong>:total</strong> visitors in total today.',
+    'beneficiary' => 'Beneficiary',
+    'volunteer_staff' => 'Volunteer / Staff',
+    'external_visitor' => 'External visitor',
 ];